Add support for wolframalpha 3.0
Allow pip to satisfy the requirement with any version between 2.0 (inclusive) and 4.0 (exclusive).

Some queries have different output when using 3.0 (vs. when using 2.x), but in most cases it
appears that the change is from returning "No text-representable result" to returning some form
of usable text, so I won't spend a ton of time trying to figure out why. It's an improvement.