release.sh: only pop stash when relevant
If nothing was stashed prior to building the release, then there's no need
to pop the stash. Popping the stash can be messy if the top entry is from
a different branch, or otherwise won't apply cleanly, and it's silly in any
case to pop a stash the release script didn't save.