Fallback to param-less W|A query if using params fails
This means newer features might not work, but it's been too long to just drop support
for the old library cold turkey. Dual support was promised for the upcoming release,
and so it will have be deprecated for the next. Then this kludge can go.

Includes documentation disclaimers that things might break with old wolframalpha lib.