Name & release sopel-wolfram v0.4.1 "Kibou ni tsuite"
Originally this release was meant to come out three weeks ago, but LuckyNumber78's
sudden passing a few days before the planned release day delayed everything.
Instead of the name "Dekopin" (lit. "forehead flick"), I chose to break from the
project's tradition a little and choose a release name in memory of my dear friend.
(LuckyNumber78 also helped me choose the release name about a year and a half ago
for sopel-wolfram v0.2.0 "Miraibashi".)

It's not stated anywhere (until now), but sopel-wolfram release names are taken
from the titles of songs by AKB48 and related groups. Ordinarily, I choose a song
title that fits what's been added, changed, or fixed in that release. The original
planned name for v0.4.1, "Dekopin", reflected the fact that the only change is a
bugfix for a bug that should never have been able to make it into a release.
(Better testing has been on my radar since day one, and is in fact issue 1 in the
GitHub issue tracker.) I deserve a dekopin for not catching it, hence the name.

The new name, "Kibou ni tsuite", goes a lot deeper than usual. Instead of just
picking a title and leaving it at that, I delved into the lyrics of several songs
to find one that really felt right as a dedication. Most of the songs listed at
Stage48.net (my go-to source for AKB discography) that deal with loss do so within
the context of a romantic breakup. That's far from matching my relationship with
LuckyNumber78, and it didn't feel right anyway to use the release name as an
expression of grief. So instead, I found a song that seemed practically written
for LuckyNumber78 to hear. 届け！