Name & release v0.5.0 "ALIVE"
As soon as I read* this song, I knew it had to be the next release name.
The title is perfect for the first release in more than two years, AND
the lyrics tie in really well on multiple levels with the basis and
inspiration of the previous release name, "Kibou ni tsuite".
(See the commit message for 3ef3c499ce4ca93161d17cceb0632b55ceba4889.)

These are powerful words. So much of the text is *painfully* relatable.
Perhaps if I overanalyzed the lyrics, I'd find that this song doesn't
fit *quite* as perfectly as the last one did. But I'm not going to do
that. I'll just use it, because it feels right.

* - Even after all this time, I still haven't actually learned enough
    Japanese to understand what I'm listening to in a song...