Vendor `wolframalpha` library
Don't know if this will be temporary or "permanent", but
jaraco/wolframalpha#35 has sat with two different PRs available to fix
it for a few months now.

Normally I don't like to vendor libraries. That said, the alternative
(requiring a pull-request branch with the fix, which could be deleted at
any time) is objectively more fragile than doing it this way.

To stay on the transparent and honorable side of upstream's MIT license,
I've included a copyright and license disclaimer (missing from the
original source file), credits to the patch author, and links.