Name & release sopel-wolfram v0.6.1 "Mirai to wa?"
This also fixes an oversight in vendoring the `wolframalpha` library:
The only apparent issue with Python 3.8 compatibility seemed to be the
lack of an `__init__.py` file in `vendor/`, when I tested it, so I did
not bump the minimum Python version up to 3.9.

`wolframalpha` upstream says it needs py3.9, but this doesn't seem to be
the case. Upstream probably just eagerly drops EOL versions.